first_test_on_board
1.Probada la fsm en el ESP8266,
2.Cambio en uint8_t por unsigned en la declaración de la estructura de los flags en door.c

+++ fsm_door_checking_test_on_board/src/door.c
@@ -4,10 +4,10 @@
 typedef union {
   struct
   {
-    uint8_t active:1;
-    uint8_t door_status:1;
-    uint8_t flag_door_open:1;
-    uint8_t pin:4;
+    unsigned active:1;
+    unsigned door_status:1;
+    unsigned flag_door_open:1;
+    unsigned pin:4;
   };
   uint8_t data;
 }flags_door_t;
